plugin/grpctrace/server.go
// Copyright 2019, OpenTelemetry Authors
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package grpctrace

import (
	"context"

	"google.golang.org/grpc"
)

func NewUnaryServerInterceptor(opts ...UnaryServerInterceptorOption) grpc.UnaryServerInterceptor {
	return func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
		res, err := handler(ctx, req)

		return res, err
	}
}

type UnaryServerInterceptorOption func(*unaryServerInterceptorConfig)

type unaryServerInterceptorConfig struct{}

func NewStreamServerInterceptor(opts ...StreamServerInterceptorOption) grpc.StreamServerInterceptor {
	return func(srv interface{}, ss grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
		return handler(srv, ss)
	}
}

type StreamServerInterceptorOption func(*streamServerInterceptorConfig)

type streamServerInterceptorConfig struct{}
